Add files via upload
My complex module and a test file for it. I ran into trouble uploading it to pyPI. I can successfully pip install the module, but the class doesn't show up for some reason. I had a similar problem yesterday which was resolved after adding the definitions to the __init__.py file, but that didn't work here